backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
alembic==1.13.2
pydantic[email]==2.9.0
reportlab==4.2.0
matplotlib==3.9.0